+++ src/main/java/com/abc/Account/Account.java
@@ -1,0 +1,65 @@
+package com.abc.Account;
+
+import com.abc.Transaction;
+
+import java.util.ArrayList;
+import java.util.List;
+
+abstract public class Account {
+
+    public List<Transaction> transactions;
+
+    public Account() {
+        this.transactions = new ArrayList<Transaction>();
+    }
+
+    public void deposit(double amount) {
+        if (amount <= 0) {
+            throw new IllegalArgumentException("amount must be greater than zero");
+        } else {
+            transactions.add(new Transaction(amount));
+        }
+    }
+
+    public void withdraw(double amount) {
+        if (amount <= 0) {
+            throw new IllegalArgumentException("amount must be greater than zero");
+        } else {
+            transactions.add(new Transaction(-amount));
+    }
+}
+
+    public abstract double interestEarned();
+//    {
+//        double amount = sumTransactions();
+//        switch(accountType){
+//            case SAVINGS:
+//                if (amount <= 1000)
+//                    return amount * 0.001;
+//                else
+//                    return 1 + (amount-1000) * 0.002;
+////            case SUPER_SAVINGS:
+////                if (amount <= 4000)
+////                    return 20;
+//            case MAXI_SAVINGS:
+//                if (amount <= 1000)
+//                    return amount * 0.02;
+//                if (amount <= 2000)
+//                    return 20 + (amount-1000) * 0.05;
+//                return 70 + (amount-2000) * 0.1;
+//            default:
+//                return amount * 0.001;
+//        }
+//    }
+
+    public double sumTransactions() {
+       return checkIfTransactionsExist(true);
+    }
+
+    private double checkIfTransactionsExist(boolean checkAll) {
+        double amount = 0.0;
+        for (Transaction t: transactions)
+            amount += t.amount;
+        return amount;
+    }
+}
