+++ src/main/java/com/abc/Account/CheckingAccount.java
@@ -1,0 +1,10 @@
+package com.abc.Account;
+
+public class CheckingAccount extends Account {
+
+    public double interestEarned() {
+        double amount = sumTransactions();
+        return 0*0.001;
+    }
+
+}
