+++ src/main/java/com/abc/Account/SavingsAccount.java
@@ -1,0 +1,12 @@
+package com.abc.Account;
+
+public class SavingsAccount extends Account {
+
+    public double interestEarned() {
+        double amount = sumTransactions();
+        if (amount <= 1000)
+            return amount * 0.001;
+        else
+            return 1 + (amount-1000) * 0.002;
+    }
+}
